Move localForIoLocal to private instances
